constants are floats

+++ FunTree.java
@@ -90,7 +90,7 @@ public class FunTree
         }
         else
         {
-            leafNode.constant = (int)(Math.random() * range - range / 2.f);
+            leafNode.constant = (float)(Math.random() * range - range / 2.f);
         }
         return leafNode;
     }
@@ -162,7 +162,7 @@ public class FunTree
         }
 
         //if leaf constant, return constant
-        return (float) current.constant;
+        return current.constant;
     }
 
     //method to make a type of computation given a string
